+++ src/pages/migrate-account.tsx
@@ -10,12 +10,11 @@ import { signIn, useSession } from "next-auth/react";
 import { useRouter } from "next/router";
 import { translations } from "../../languages/translations";
 import { api } from "../utils/api";
-import { z } from "zod";
 import AlreadySignedInPage from "../Components/AlreadySignedInPage";
 import InputField from "../Components/InputField";
 import Loading from "../Components/Loading";
 import LoadingPage from "../Components/LoadingPage";
-import { migrationValidationObject } from "../server/api/routers/migrate";
+import { migrationValidationObject } from "../validation/validation";
 /**
  * @returns {JSX.Element} - React component.
  * @description Page for migrating an old applicant account to the new system.
@@ -39,13 +38,13 @@ export default function MigrateAccount() {
   const migrateAccount = api.migration.createUserForOldApplication.useMutation();
 
   /* Validation */
-  const inputValidation = migrationValidationObject.safeParse(newUser);
+  const validation = migrationValidationObject.safeParse(newUser);
   const isFieldValid = (field: string) =>
-    inputValidation.success ? true : inputValidation.error.issues.find((i) => i.path[0] === field) === undefined;
+    validation.success ? true : validation.error.issues.find((i) => i.path[0] === field) === undefined;
 
   /* Handelers */
   const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-    if (!inputValidation.success) return;
+    if (!validation.success) return;
 
     e.preventDefault();
     setLoading(true);
@@ -80,25 +79,9 @@ export default function MigrateAccount() {
       {/* Input fields */}
       <form onSubmit={handleSubmit}>
         <div className="grid gap-20 mb-6 md:grid-cols-3">
-          {InputField("email", "Email", "text", newUser.email, true, handleUpdateNewUser, isFieldValid("email"))}
-          {InputField(
-            "username",
-            "Username",
-            "text",
-            newUser.username,
-            true,
-            handleUpdateNewUser,
-            isFieldValid("username"),
-          )}
-          {InputField(
-            "password",
-            "Password",
-            "password",
-            newUser.password,
-            true,
-            handleUpdateNewUser,
-            isFieldValid("password"),
-          )}
+          {InputField("email", "Email", "text", newUser.email, true, handleUpdateNewUser, isFieldValid)}
+          {InputField("username", "Username", "text", newUser.username, true, handleUpdateNewUser, isFieldValid)}
+          {InputField("password", "Password", "password", newUser.password, true, handleUpdateNewUser, isFieldValid)}
         </div>
 
         {/* Submit button */}
@@ -108,11 +91,11 @@ export default function MigrateAccount() {
           ) : (
             <input
               className={`text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800${
-                inputValidation.success ? "" : " opacity-50 cursor-not-allowed"
+                validation.success ? "" : " opacity-50 cursor-not-allowed"
               }`}
               type="submit"
               value={"Migrate"}
-              disabled={!inputValidation.success}
+              disabled={!validation.success}
             />
           )}
         </div>
